silence signed/unsigned comparision warning

+++ lib/dns/dst_api.c
@@ -18,7 +18,7 @@
 
 /*
  * Principal Author: Brian Wellington
- * $Id: dst_api.c,v 1.1.6.14 2009/09/24 22:25:30 marka Exp $
+ * $Id: dst_api.c,v 1.1.6.15 2009/09/25 01:48:28 marka Exp $
  */
 
 /*! \file */
@@ -1224,7 +1224,7 @@ addsuffix(char *filename, unsigned int len, const char *ofilename,
 	n = snprintf(filename, len, "%.*s%s", olen, ofilename, suffix);
 	if (n < 0)
 		return (ISC_R_FAILURE);
-	if (n >= len)
+	if ((unsigned int)n >= len)
 		return (ISC_R_NOSPACE);
 	return (ISC_R_SUCCESS);
 }
